Create a Seurat object with Splice object

Create a Seurat and Splice object from the output file from saveExonList().

Usage

createSpliceSeurat(
  path,
  project = "SeuratProject",
  assay = "RNA",
  min.cells = 3,
  min.features = 20,
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  ...
)

Arguments

path

The path to store the output files from saveExonList()

project

The name of the project for this object

assay

The assay name for the count matrix

min.cells

The minimum number of cells to express a gene

min.features

The minimum number of genes cell to express

names.field

For the initial identity class for each cell, choose this field from the cell's name.

names.delim

For the initial identity class for each cell, choose this delimiter from the cell's column name

meta.data

Additional cell-level metadata to add to the Seurat object.

...

other parameters for Seurat::CreateSeuratObject

Details

The Seraut object is created to store the gene expression information, and the Splice object for isoform information. The Splice object is embedded in the misc slot in Seurat object