generic genes_groups_GLRT function for Seurat object

an interface function to run GLRT for all exons in multiple genes for multiple cell group pairs from a Splice object.

Usage

# S4 method for Seurat,character,character,character
genes_groups_GLRT(
  object,
  genes,
  group1s,
  group2s,
  group_col = NULL,
  slot = "splice",
  ...
)

Arguments

object

the Seurat object

genes

a vector of gene names to do GLRT

group1s

group names or cell names

group2s

group names or cell names

group_col

if group_col is NULL, group1s and group2s should be cell names in each group. If group_col is specified as a colname in Seurat cell meta data, group1s and group2s should be group names.

slot

the slot name to store Splice object in Seurat object

...

parameters for genes_groups_GLRT.base

Value

a dataframe of the test result. Each line is an exon with p-values and parameters for the psi distributions in two cell populations if they have significant difference