genes_multigroups_GLRT

Do the generalized likelihood ratio test for multiples gene across multiple cell groups.

Usage

genes_multigroups_GLRT(
  spliceOb,
  genes,
  meta,
  cell_col = "cell",
  group_col = "group",
  filter = TRUE,
  q_thresh = 0.05,
  ...
)

Arguments

spliceOb

the Splice object

genes

target genes to be tested

meta

a dataframe recording the meta data for each, should have at lease two columns, one indicate the cell barcode and the other one indicate the cell group information.

cell_col, group_col

The column names in the meta to record the cell/group info.

filter

A bool flag to indicate if insignificant results should be filtered out.

q_thresh

The threshold for the FDR controled q value.

...

Other parameters used in gene_GLRT.base.

Value

a dataframe of the test result. Each line is an exon with p-values and parameters for the psi distributions in two cell populations if they have significant difference